requirements.txt
groq
python-dotenv
google-auth
google-auth-oauthlib
google-api-python-client